requirements.txt

Flask-Cors==4.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
email-validator==2.1.0
WTForms==3.1.1
cryptography==41.0.7